Added offline caching of project.json and subsequent requests
When calling Formio.cacheOfflineProject, it will store the provided project.json in localStorage. Then when called again later, it will just pull the stored cache in localStorage.

Offline cache is separate from regular cache. When a form GET/INDEX request is made while offline, then it is served from the offline cache if available. If a form is requested while online, and that form's project has been cached, the result of the request is used to update the offline cache.